+++ ViewModels/Areas/Administrator/Inventoryamount/InventoryamountViewModel.cs
@@ -13,7 +13,7 @@ namespace ViewModels.Areas.Administrator.Inventoryamount
             (ResourceType = typeof(Resources.Model.Inventoryamount),
             Name = Resources.Model.Strings.InventoryamountKeys.Inventorytonnage)]
         #endregion
-        public int Inventorytonnage { get; set; }
+        public double Inventorytonnage { get; set; }
         #region ProductName
         [System.ComponentModel.DataAnnotations.Display
             (ResourceType = typeof(Resources.Model.Inventoryamount),
